# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para sistema de pesagem de mirtilo
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Worker {
  id: string;
  name: string;
  active: boolean;
  createdAt: string;
}

export interface CreateWorkerBody {
  /** Worker identifier (e.g. W001) */
  id: string;
  name: string;
}

export interface UpdateWorkerBody {
  name?: string;
  active?: boolean;
}

export type QualityIssue = (typeof QualityIssue)[keyof typeof QualityIssue];

export const QualityIssue = {
  CALIBRE: "CALIBRE",
  PENDUNCULOS: "PENDUNCULOS",
  VERDE: "VERDE",
  MOLE: "MOLE",
  OUTROS: "OUTROS",
} as const;

export interface WeighRecord {
  id: number;
  workerId: string;
  workerName: string;
  timestamp: string;
  weightGrams: number;
  unit: string;
  scaleId: string;
  rawLine: string;
  syncStatus: string;
  createdAt: string;
  /** @nullable */
  editedAt?: string | null;
  qualityIssues: QualityIssue[];
}

export interface CreateWeighRecordBody {
  workerId: string;
  timestamp?: string;
  weightGrams: number;
  unit: string;
  scaleId: string;
  rawLine: string;
  qualityIssues?: QualityIssue[];
}

export interface UpdateWeighRecordBody {
  /**
   * @minimum 50
   * @maximum 10000
   */
  weightGrams: number;
  qualityIssues?: QualityIssue[];
}

/**
 * Per-type counts of quality occurrences
 */
export interface QualityIssueCounts {
  CALIBRE: number;
  PENDUNCULOS: number;
  VERDE: number;
  MOLE: number;
  OUTROS: number;
}

export interface WorkerDailyStats {
  workerId: string;
  workerName: string;
  totalCaixas: number;
  totalKg: number;
  mediaGrPorCaixa: number;
  caixasPorHora: number;
  /**
   * Hours from check-in to check-out (or now, if still working). Null if no attendance record.
   * @nullable
   */
  hoursWorked: number | null;
  /**
   * totalKg / hoursWorked. Null if hoursWorked is null or zero.
   * @nullable
   */
  kgPorHora: number | null;
  /** @nullable */
  primeiroRegisto: string | null;
  /** @nullable */
  ultimoRegisto: string | null;
  rankKg: number;
  /** Total number of quality flags reported across all boxes for this worker. */
  totalIssues: number;
  issuesByType: QualityIssueCounts;
}

/**
 * A single entrada/saída (shift) within a day. A day can have several.
 */
export interface WorkerTimesheetShift {
  id: number;
  /** @nullable */
  checkInAt: string | null;
  /** @nullable */
  checkOutAt: string | null;
  /** @nullable */
  hoursWorked: number | null;
  /**
   * hoursWorked * hourlyRate for this shift (null if either is missing)
   * @nullable
   */
  pay: number | null;
}

export interface WorkerTimesheetDay {
  date: string;
  /**
   * First check-in of the day (aggregate over all shifts).
   * @nullable
   */
  checkInAt: string | null;
  /**
   * Last check-out of the day, or null if any shift is still open.
   * @nullable
   */
  checkOutAt: string | null;
  /**
   * Sum of hoursWorked across the day's closed shifts.
   * @nullable
   */
  hoursWorked: number | null;
  /**
   * hoursWorked * hourlyRate (null if either is missing)
   * @nullable
   */
  pay: number | null;
  /** Number of quality flags reported on this day. */
  totalIssues: number;
  issuesByType: QualityIssueCounts;
  shifts: WorkerTimesheetShift[];
}

export interface WorkerTimesheet {
  workerId: string;
  workerName: string;
  from: string;
  to: string;
  /** @nullable */
  hourlyRate: number | null;
  days: WorkerTimesheetDay[];
  /** Number of days with at least a check-in */
  totalDays: number;
  totalHours: number;
  /** @nullable */
  totalPay: number | null;
}

export interface AttendanceEntry {
  workerId: string;
  workerName: string;
  /** ISO date YYYY-MM-DD */
  date: string;
  /** @nullable */
  checkInAt: string | null;
  /** @nullable */
  checkOutAt: string | null;
  /**
   * Hours between check-in and check-out (null if not yet checked out)
   * @nullable
   */
  hoursWorked: number | null;
}

export interface AttendanceMutationBody {
  workerId: string;
  /** ISO date YYYY-MM-DD (defaults to today) */
  date?: string;
}

export interface AttendanceBulkBody {
  /** ISO date YYYY-MM-DD (defaults to today) */
  date?: string;
  /** Restrict the bulk action to these worker IDs (defaults to all active workers) */
  workerIds?: string[];
}

export interface ShiftCreateBody {
  workerId: string;
  /** ISO date YYYY-MM-DD */
  date: string;
  checkInAt: string;
  /** @nullable */
  checkOutAt?: string | null;
}

/**
 * All fields optional; omit a field to leave it unchanged.
 */
export interface ShiftUpdateBody {
  /** Move the shift to this ISO date YYYY-MM-DD */
  date?: string;
  checkInAt?: string;
  /** @nullable */
  checkOutAt?: string | null;
}

export interface DailyReport {
  date: string;
  workers: WorkerDailyStats[];
  totalRecords: number;
  totalKg: number;
}

export type ListWeighRecordsParams = {
  workerId?: string;
  /**
   * ISO date string YYYY-MM-DD
   */
  date?: string;
  limit?: number;
};

export type ListAttendanceParams = {
  /**
   * ISO date YYYY-MM-DD (defaults to today)
   */
  date?: string;
};

export type GetWorkerTimesheetParams = {
  /**
   * ISO date YYYY-MM-DD (default 30 days ago)
   */
  from?: string;
  /**
   * ISO date YYYY-MM-DD (default today)
   */
  to?: string;
  /**
   * Optional hourly rate for pay calculation
   */
  hourlyRate?: number;
};

export type ExportWorkerTimesheetParams = {
  from?: string;
  to?: string;
  hourlyRate?: number;
};

export type GetDailyReportParams = {
  /**
   * ISO date YYYY-MM-DD (defaults to today)
   */
  date?: string;
};

export type ExportRecordsParams = {
  /**
   * ISO date YYYY-MM-DD (defaults to today)
   */
  date?: string;
  workerId?: string;
};