# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: API para sistema de pesagem de mirtilo
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List worker attendance for a given day (defaults to today)
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
          description: ISO date YYYY-MM-DD (defaults to today)
      responses:
        "200":
          description: Attendance list with one entry per active worker
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceEntry"
  /attendance/check-in:
    post:
      operationId: checkInWorker
      tags: [attendance]
      summary: Check in a single worker for the day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceMutationBody"
      responses:
        "200":
          description: Attendance entry for the worker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Worker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/check-out:
    post:
      operationId: checkOutWorker
      tags: [attendance]
      summary: Check out a single worker for the day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceMutationBody"
      responses:
        "200":
          description: Attendance entry for the worker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Worker not checked in today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/check-in-all:
    post:
      operationId: checkInAll
      tags: [attendance]
      summary: Check in all active workers for the day
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceBulkBody"
      responses:
        "200":
          description: Updated attendance list for the day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceEntry"
  /attendance/check-out-all:
    post:
      operationId: checkOutAll
      tags: [attendance]
      summary: Check out all currently checked-in workers for the day
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceBulkBody"
      responses:
        "200":
          description: Updated attendance list for the day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceEntry"
  /attendance/shift:
    post:
      operationId: createAttendanceShift
      tags: [attendance]
      summary: Manually create a shift (entrada/saída) for a worker on a day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftCreateBody"
      responses:
        "201":
          description: The created shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Worker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Shift conflict (overlap or open shift already exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/shift/{id}:
    patch:
      operationId: updateAttendanceShift
      tags: [attendance]
      summary: Edit a shift's times or move it to another day
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftUpdateBody"
      responses:
        "200":
          description: The updated shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Shift not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Shift conflict (overlap or open shift already exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAttendanceShift
      tags: [attendance]
      summary: Delete a shift by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Shift deleted
        "404":
          description: Shift not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AttendanceEntry:
      type: object
      properties:
        workerId:
          type: string
        workerName:
          type: string
        date:
          type: string
          description: ISO date YYYY-MM-DD
        checkInAt:
          type: ["string", "null"]
          format: date-time
        checkOutAt:
          type: ["string", "null"]
          format: date-time
        hoursWorked:
          type: ["number", "null"]
          description: Hours between check-in and check-out (null if not yet checked out)
      required:
        - workerId
        - workerName
        - date
        - checkInAt
        - checkOutAt
        - hoursWorked
    AttendanceMutationBody:
      type: object
      properties:
        workerId:
          type: string
        date:
          type: string
          description: ISO date YYYY-MM-DD (defaults to today)
      required:
        - workerId
    AttendanceBulkBody:
      type: object
      properties:
        date:
          type: string
          description: ISO date YYYY-MM-DD (defaults to today)
        workerIds:
          type: array
          description: Restrict the bulk action to these worker IDs (defaults to all active workers)
          items:
            type: string
    ShiftCreateBody:
      type: object
      properties:
        workerId:
          type: string
        date:
          type: string
          description: ISO date YYYY-MM-DD
        checkInAt:
          type: string
          format: date-time
        checkOutAt:
          type: ["string", "null"]
          format: date-time
      required:
        - workerId
        - date
        - checkInAt
    ShiftUpdateBody:
      type: object
      description: All fields optional; omit a field to leave it unchanged.
      properties:
        date:
          type: string
          description: Move the shift to this ISO date YYYY-MM-DD
        checkInAt:
          type: string
          format: date-time
        checkOutAt:
          type: ["string", "null"]
          format: date-time
...