# lib/api-zod/src/generated/types/shiftCreateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para sistema de pesagem de mirtilo
 * OpenAPI spec version: 1.0.0
 */

export interface ShiftCreateBody {
  workerId: string;
  /** ISO date YYYY-MM-DD */
  date: string;
  checkInAt: Date;
  /** @nullable */
  checkOutAt?: Date | null;
}