# lib/api-zod/src/generated/types/shiftUpdateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para sistema de pesagem de mirtilo
 * OpenAPI spec version: 1.0.0
 */

/**
 * All fields optional; omit a field to leave it unchanged.
 */
export interface ShiftUpdateBody {
  /** Move the shift to this ISO date YYYY-MM-DD */
  date?: string;
  checkInAt?: Date;
  /** @nullable */
  checkOutAt?: Date | null;
}